+++ CHANGELOG.md
@@ -4,6 +4,16 @@ This file contains the changes made to the package.
 
 The sections are in descending order of the change date.
 
+## [4.1.0] - 2026-06-09
+
+### Added
+
+- Support for arrays at the root level for decorators instead of only objects. e.g. `@QueryWithZod(Task.array())` instead of `@Query(() => [Task])`
+
+### Changed
+
+- Fixed `@ZodArgs` not outputting the default value in the generated GraphQL schema.
+
 ## [4.0.1] - 2026-06-05
 
 ### Changed